requirements.txt
lark-parser==0.7.8
PyQt5==5.13.2
PyQt5-sip==12.7.0
python-statemachine==0.7.1
